directly define and export context related types
