--- packages/core/src/context/ContextChunk.ts
@@ -1,16 +1,0 @@
-import { ContextSource } from "./ContextSource.js";
-
-export type ContextChunk = {
-  id: string;
-  source: ContextSource;
-  content: string;
-  metadata?: {
-    symbols?: string[];
-    language?: string;
-    path?: string;
-    lineStart?: number;
-    lineEnd?: number;
-    kind?: "code" | "test" | "doc" | "config";
-    [key: string]: unknown;
-  };
-};
