--- packages/core/src/context/RetrievedContext.ts
@@ -1,9 +1,0 @@
-import type { ContextChunk } from "./ContextChunk.js";
-
-export type RetrievedContext = {
-  chunks: ContextChunk[];
-  stats: {
-    totalChunks: number;
-    truncated: boolean;
-  };
-};
